Refactored Makefile and Added An Assembly (NASM) Boot Stub
There Are Many More Improvements I Could Probably Make to The Makefile But I'm Not Really That Good at it. I Also Added An Assembly (NASM) File And Made it The Entry Point, Right Now it Does Nothing Oter Than Call The Main Kernel Function, But I Heard It's A Good Idea to Have An Assembly Boot Stub so Maybe it Will Come in Handy Later.